requirements.txt
matplotlib
Pillow
PyMuPDF
PySide6
PyYAML
shapely
svgelements
scipy
Hershey-Fonts
